添加组织架构逻辑（测试排序树通用方法）

+++ zhuma-commons/zhuma-commons-public/src/main/java/com/zm/zhuma/commons/model/po/SortTreePO.java
@@ -2,9 +2,9 @@ package com.zm.zhuma.commons.model.po;
 
 public interface SortTreePO<PK> extends TreePO<PK>, Comparable<SortTreePO> {
 
-	Long getSort();
+	Integer getSort();
 
-	void setSort(Long sort);
+	void setSort(Integer sort);
 
 	@Override
 	default int compareTo(SortTreePO sortTree) {
@@ -12,7 +12,7 @@ public interface SortTreePO<PK> extends TreePO<PK>, Comparable<SortTreePO> {
 			return -1;
 		}
 
-		return Long.compare(getSort() == null ? 0 : getSort(), sortTree.getSort() == null ? 0 : sortTree.getSort());
+		return Integer.compare(getSort() == null ? 0 : getSort(), sortTree.getSort() == null ? 0 : sortTree.getSort());
 	}
 	
 }
